임우정 : 화살생성 ,몬스터 데미지 구현중

+++ GameEngineContents/GlobalContentsValue.h
@@ -14,7 +14,7 @@ typedef struct UnitInfo
 	int m_MaxHp;
 	int Gold;
 	int ArrowCount;
-	float Dammage;
+	int Dammage;
 	WEAPONTYPE Weapontype;
 
 }UINFO;
